+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "self-assessment"
-version = "1.1.2"
+version = "2.0.0"
 edition = "2021"
 description = "A CLI tool that generates a list of pull requests raised and reviewed in the Guardian's GitHub organisation, as well as an optional summary of the user's Trello boards and cards."
 license = "MIT"
@@ -24,6 +24,7 @@ rust-ini = "0.17"
 handlebars = "4.2.1"
 colorsys = "0.6.5"
 chrono = { version = "0.4", features = ["serde"] }
+anyhow = "1.0.66"
 
 [[bin]]
 name = "self-assessment"
