+++ src/models.rs
@@ -100,12 +100,12 @@ pub struct Milestone {
     pub number: u32,
     pub state: String,
     pub title: String,
-    pub description: String,
+    pub description: Option<String>,
     pub creator: User,
     pub open_issues: u32,
     pub closed_issues: u32,
     pub created_at: String,
-    pub updated_at: String,
+    pub updated_at: Option<String>,
     pub closed_at: Option<String>,
     pub due_on: Option<String>,
 }
